Move TransactionFamily to parent module
Move the TransactionFamily to the parent module, to remove potential circular references
